requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

matplotlib==3.10.0
mitreattack-python==3.0.8
numpy==2.2.2
pandas==2.2.3
scikit-learn==1.6.1
